filter out allowed zones list if TD Enhancement Pack mod is used
It allows disabling showing an area for colonists, animals or mechs,
and it's easy to get to 10+ zones that way. Which gets hard to use
with Hospitality showing the full list, twice.

So do not show areas disabled for colonists. As a special feature,
allow those that have name ending in "Hospitality", so that it's
possible to disable showing them everywhere else.